setup LeastRecentlyUsed
