+++ src/least-recently-used/LeastRecentlyUsed.test.ts
@@ -1,0 +1,7 @@
+import { LinkedList } from "../index.js";
+
+describe("LeastRecentlyUsed", () => {
+    it("", () => {
+
+    });
+});
