+++ src/least-recently-used/LeastRecentlyUsed.ts
@@ -1,0 +1,5 @@
+export class LeastRecentlyUsed {
+    constructor() {
+
+    }
+}
