+++ src/least-recently-used/README.md
@@ -1,0 +1,3 @@
+# LeastRecentlyUsed
+
+An implementation of a `LinkedList` that has a set size and adds new nodes or moves existing nodes to the front of the list while drpping old nodes. It is useful for caching data that will probably be needed more than once somewhat consecutively.
