+++ src/least-recently-used/index.ts
@@ -1,0 +1,1 @@
+export * from "./LeastRecentlyUsed.js";
