+++ src/node/README.md
@@ -4,5 +4,5 @@ Composable properties to build Nodes for each data-structure.
 
 ## List
 
-- [Completable](index.ts): Initializes a boolean state.
+- [Completable](Completable.ts): Initializes a boolean state.
 - [Countable](Countable.ts): Initializes a counter that can be incremented and decremented.
